chore : 테스트 파일 삭제 및 폴더 정리
