--- src/mocks/recipe.ts
@@ -1,1 +1,0 @@
-export const ExampleRecipeData = {};
