--- src/types/exampleTypes.ts
@@ -1,1 +1,0 @@
-// 타입정의
