--- src/utils/exampleUtils.ts
@@ -1,1 +1,0 @@
-// 유틸함수 정의
